requirements.txt
alembic==0.9.8
click==6.7
elasticsearch==6.2.0
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.2
pandas==0.22.0
progressbar2==3.37.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
python-editor==1.0.3
python-utils==2.3.0
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.4
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
